backend/requirements.txt
# Web 框架和服务器
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
starlette>=0.14.0,<0.15.0

# 数据库
sqlalchemy>=1.4.0,<2.0.0
pymysql>=1.0.2,<2.0.0
mysql-connector-python>=8.0.27,<9.0.0
alembic>=1.7.0,<2.0.0
cryptography>=35.0.0,<36.0.0
dbutils>=3.0.0,<4.0.0
peewee>=3.14.0,<4.0.0

# 认证和安全
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
PyJWT>=2.3.0,<3.0.0

# 工具
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<3.0.0
aiofiles>=0.7.0,<0.8.0
email-validator>=1.1.3,<2.0.0
loguru>=0.6.0,<0.7.0
Pillow>=9.0.0,<10.0.0

# 日期时间处理
pytz>=2021.3,<2022.0 